+++ CHANGELOG.md
@@ -21,5 +21,6 @@
 - community: noice UI replacement for messages, popups & command line
 - mapping: add `SPC W` save as key binding prompting for file name
 - snippets: luasnip plugin and json format snippets
+- practicalli: `fd` binding for `ESC` and startup banner
 
 ## Changed